Heart Rate and Map Demo WatchKit App
This app is not intended to be used as an everyday, practical app. It was created for the sole purpose of demonstrating some of the Apple Watch's technologies and how to implement them. Of course, these implementations can be used as a foundation for more functional applications.

This project demonstrates the capabilities of the GPS and heart rate sensor of the Apple Watch.  The Watch app monitors the user's heart rate current location, and this data is sent to the corresponding iPhone app.  Use the iPhone app to view this data and to compare the iPhone GPS to the Watch GPS on a map.  You can also use the map to place down multiple location pins.  Once the user is in close proximity to the pin (according to the GPS), a notification will fire off saying so.